Update Name and Attribute
Renamed to multiMapFSC, and replaced geolocation with coordinates to reduce confusion.
Did not replace mapFSC aura component, as that is a bit simpler to use for single marker scenario.
Delimiter is left as ';' instead of ',' due to the fact that address attributes may have ',' in the string.  Coordinates will not, so could update that delimiter, but left as this could cause confusion in forming the input.