Update ExecuteSOQLTest.cls
I attempted to push the ExecuteSOQL class and this test class to production and got test class failures. One of the developers on my team fixed it so that it would pass. I'm not sure exactly how the changes he made impact overall though I wanted to offer them for the next person running into this issue. Thanks for making these great tools for admins.