Fix issue with commas not being replace after parse
Fix for issue #902 reported here.
https://github.com/alexed1/LightningFlowComponents/issues/902.
The merged code was missing calls to add commas and quotes back in after
the values have been parsed.
The tests were missing asserts to validate that the final parsed value
matches what we expect when the value contains commas, newlines and
quotes. Added asserts to verify that.