Add a new input param for specifying date format
If the date is in 'yyyy-mm-dd' format, the date format input is ignored.
Else, the user can specify if the date is in one of 'MM/dd/yyyy'  or
'dd/MM/yyyy' formats. If the date is not in 'yyyy-MM-dd' format and the
date format is not specified, date is attempted to be parsed in the
user's local date format. Added more tests for percent and date inputs.