Fix issue with External Objects date fields
Salesforce has a weird behavior with External Objects where date fields are formatted as datetime fields (they include the time in the string representation). This was causing the values to be processed incorrectly here. `Date.parse()` was returning `NaN` and then `record[date]` was being overwritten with `null` as a result. 
By checking for `isNaN` we can prevent this from happening.